# sve2 histcnt support for fast ondemand parsing